DE-3484 Moved utils.py back into utils folder and break logic out into separate files to avoid cyclic dependency
