+++ __tests__/integration/test_datetime_parser.py
@@ -4,6 +4,7 @@ from task_script_utils.datetime_parser.parser import (
     parse,
     DatetimeConfig,
 )
+from task_script_utils.datetime_parser.utils.parsing import _parse_with_formats
 
 from task_script_utils.datetime_parser.parser_exceptions import DatetimeParserError
 
@@ -53,6 +54,13 @@ parse_with_no_datetime_formats_test_cases = {
     "Sunday, May 26th 2013 12:12:12 AM ZST Asia/Kolkata": None,
     # Missing Month
     "Sunday, 26 2013 12:12:12 AM Asia/Kolkata": None,
+}
+
+format_list_with_no_tz_dict_test_cases = {
+    "Sunday, May 26th 2013 12:12:12 AM IST Asia/Kolkata": None,
+    "Sunday, May 26th 2013 12:12:12 AM ZST Asia/Kolkata": None,
+    "Sunday, May 26th 2013 12:12:12 AM Asia/Kolkata": "2013-05-26T00:12:12+05:30",
+    "Sunday, May 26th 2013 12:12:12 AM": None,
 }
 
 dateutil_parser_test_cases = {
