+++ __tests__/unit/test_check_for_mutual_ambiguity.py
@@ -1,7 +1,8 @@
 import pytest
 from pendulum import datetime as pendulum_datetime
 from task_script_utils.datetime_parser import DatetimeConfig
-from task_script_utils.datetime_parser.parser import parse, parse_with_formats
+from task_script_utils.datetime_parser.parser import parse
+from task_script_utils.datetime_parser.utils.parsing import parse_with_formats
 from task_script_utils.datetime_parser.parser_exceptions import (
     AmbiguousDatetimeFormatsError,
 )
