+++ __tests__/unit/test_datetime_parser_utils.py
@@ -1,8 +1,9 @@
 # pylint: disable=C0114
 # pylint: disable=E0401
 import pytest
-from task_script_utils.datetime_parser.utils import parse_with_formats
+from task_script_utils.datetime_parser.utils.parsing import parse_with_formats
 from task_script_utils.datetime_parser import DatetimeConfig
+from task_script_utils.datetime_parser.parser_exceptions import DatetimeParserError
 
 format_list_with_no_tz_dict_test_cases = {
     "Sunday, May 26th 2013 12:12:12 AM IST Asia/Kolkata": None,
@@ -27,9 +28,13 @@ formats_list = [
 )
 def test_parse_with_formats_with_no_tz_dict(input_, expected):
     """Test `parse_with_formats` when no tz_dict is present in datetime_config"""
-    parsed_datetime, _ = parse_with_formats(input_, DatetimeConfig(), formats_list)
-
-    if parsed_datetime is None:
-        assert parsed_datetime == expected
+    if expected is None:
+        with pytest.raises(DatetimeParserError):
+            parse_with_formats(input_, formats_list, DatetimeConfig())
     else:
-        assert parsed_datetime.isoformat() == expected
+        parsed_datetime = parse_with_formats(input_, formats_list, DatetimeConfig())
+
+        if parsed_datetime is None:
+            assert parsed_datetime == expected
+        else:
+            assert parsed_datetime.isoformat() == expected
